Move package into AirGap module and update imports
Move project files into an AirGap/ package and update references accordingly. Renamed/relocated modules and resources into AirGap/*, updated imports to use the AirGap package (e.g. from AirGap.lib... / AirGap.commands...), adjusted the GitHub workflow to read AirGap/AirGap.manifest and AirGap/config.py, and updated ruff.toml per-file-ignores to the new paths. No functional logic changes aside from path/import fixes.
