+++ AirGap/lib/export_manager.py
@@ -5,7 +5,7 @@ import adsk.cam
 import adsk.core
 import adsk.fusion
 
-from lib.audit_logger import AuditLogger
+from AirGap.lib.audit_logger import AuditLogger
 
 
 class LocalExportManager:
