+++ AirGap/lib/offline_enforcer.py
@@ -3,9 +3,9 @@ import threading
 
 import adsk.core
 
-import config
-from lib.audit_logger import AuditLogger
-from lib.session_manager import ITARSessionManager
+import AirGap.config as config
+from AirGap.lib.audit_logger import AuditLogger
+from AirGap.lib.session_manager import ITARSessionManager
 
 
 class OnlineStatusChangedHandler(adsk.core.ApplicationEventHandler):
