+++ AirGap/lib/session_manager.py
